developments_build/python/requirements.txt
asyncio==3.4.3
backcall==0.1.0
bleach==3.1.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.3
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.0
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
nest-asyncio==0.9.10
notebook==5.7.4
numpy==1.16.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.5.0
prompt-toolkit==2.0.7
psycopg2==2.7.7
ptyprocess==0.6.0
py4j==0.10.7
Pygments==2.3.1
python-dateutil==2.7.5
python-geosupport==1.0.3
pytz==2018.9
pyzmq==17.1.2
qtconsole==4.4.3
Send2Trash==1.5.0
six==1.12.0
SQLAlchemy==1.2.16
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
uvloop==0.12.0
wcwidth==0.1.7
